+++ frontend/src/components/contextSelectors/AssignmentSelectionMenu.tsx
@@ -1,8 +1,8 @@
 import { ReactElement, useEffect, useState } from "react";
 import { Assignment, PaletteAPIResponse } from "palette-types";
 import { useFetch } from "@hooks";
-import { useCourse } from "../context";
-import { useAssignment } from "../context/AssignmentProvider.tsx";
+import { useCourse } from "@context";
+import { useAssignment } from "../../context/AssignmentProvider.tsx";
 import { LoadingDots } from "@components";
 
 export default function AssignmentSelectionMenu({
