+++ frontend/src/components/layout/Footer.tsx
