+++ frontend/src/components/layout/MainPageTemplate.tsx
