+++ frontend/src/components/layout/Navbar.tsx
@@ -3,11 +3,11 @@ import { useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import MenuItem from "@mui/material/MenuItem";
 import Menu from "@mui/material/Menu";
-import ActiveCourseSelection from "./ActiveCourseSelection.tsx";
-import { Dialog } from "./Dialog.tsx";
-import CourseSelectionMenu from "./CourseSelectionMenu.tsx";
-import ActiveAssignmentSelection from "./ActiveAssignmentSelection.tsx";
-import AssignmentSelectionMenu from "./AssignmentSelectionMenu.tsx";
+import ActiveCourseSelection from "../contextSelectors/ActiveCourseSelection.tsx";
+import { Dialog } from "../modals/Dialog.tsx";
+import CourseSelectionMenu from "../contextSelectors/CourseSelectionMenu.tsx";
+import ActiveAssignmentSelection from "../contextSelectors/ActiveAssignmentSelection.tsx";
+import AssignmentSelectionMenu from "../contextSelectors/AssignmentSelectionMenu.tsx";
 
 function Navbar() {
   const [userAnchor, setUserAnchor] = useState<null | HTMLElement>(null);
