+++ frontend/src/main.tsx
@@ -2,18 +2,18 @@
 Entry point for the entire application.
  */
 
-import { StrictMode } from "react";
-import { createRoot } from "react-dom/client";
+import {StrictMode} from "react";
+import {createRoot} from "react-dom/client";
 
 import "./index.css";
 
-import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import {BrowserRouter as Router, Route, Routes} from "react-router-dom";
 
-import NotFoundPage from "./components/NotFoundPage";
+import NotFoundPage from "./features/errorPages/NotFoundPage.tsx";
 
-import { CourseProvider } from "./context/CourseProvider";
-import { AssignmentProvider } from "./context/AssignmentProvider.tsx";
-import { GradingMain, Home, RubricBuilderMain, SettingsMain } from "@features";
+import {CourseProvider} from "./context/CourseProvider";
+import {AssignmentProvider} from "./context/AssignmentProvider.tsx";
+import {GradingMain, Home, RubricBuilderMain, SettingsMain} from "@features";
 
 // Defined a "root" div in index.html that we pull in here and then call the React render method.
 createRoot(document.getElementById("root")!).render(
